test(Effect): fold track mapped-type coverage into existing suites
Move the typetest assertions into packages/effect/typetest/Effect.tst.ts and the
runtime preservation tests into the existing describe("track") block in
packages/effect/test/Metric.test.ts, using that file's nextId() helper instead of
hardcoded metric ids. No coverage change: 2288 typetests / 5590 assertions before
and after.
